Enhance benchmarks and query execution error handling
- Updated CMakeLists.txt to conditionally include LLM inference benchmarks based on THEMIS_ENABLE_LLM.
- Refactored error handling in bench_query.cpp to use a unified result type for query execution, improving clarity and consistency.
- Modified bench_rag_ethics.cpp to use fully qualified names for RetrievedDocument, ensuring clarity in the context of RAG ethics.
- Improved CMake configuration for OpenSSL in Dependencies.cmake to allow for optional inclusion and provide warnings if not found.
- Enhanced inmemory_cache_impl.h to use basic_regex directly, avoiding potential macro conflicts.
- Added adaptive query optimization components in query_optimizer.h, including AdaptiveQueryStats, AdaptivePlanSelector, DistributedQueryCostModel, and MultiIndexOptimizer.
- Updated cross_shard_transaction.h to include <set> for cycle detection logic.
- Refactored raft_state.h to simplify access to commit_index without using atomic load.
- Cleaned up query_api_handler.cpp to improve error handling for query execution results.
- Added a build script for Windows to streamline the build process with Visual Studio and CMake.
- Updated tests to use in-memory RocksDB for SecondaryIndexManager initialization, ensuring isolation and speed in unit tests.
- Adjusted tracing tests to use double for numeric attributes, ensuring consistency in attribute types.

+++ tests/test_distributed_tracing.cpp
@@ -143,7 +143,7 @@ TEST_F(DistributedTracingTest, ConcurrentSpans) {
         threads.emplace_back([spans_per_thread]() {
             for (int j = 0; j < spans_per_thread; j++) {
                 ScopedSpan span("concurrent.operation");
-                span.setAttribute("thread_local", j);
+                span.setAttribute("thread_local", static_cast<double>(j));
             }
         });
     }
